+++ ocs_ci/ocs/monitoring.py
@@ -4,7 +4,7 @@ import json
 
 from ocs_ci.ocs.ocp import OCP
 from ocs_ci.utility import templating
-from ocs_ci.ocs import constants, defaults, ocp
+from ocs_ci.ocs import constants, defaults
 from ocs_ci.ocs.resources.pvc import get_all_pvcs, PVC
 from ocs_ci.ocs.resources.pod import get_pod_obj
 from tests import helpers
@@ -214,15 +214,3 @@ def prometheus_health_check(name=constants.MONITORING, kind=constants.CLUSTER_OP
 
     logging.error(f"Prometheus cluster is degraded {health_conditions}")
     return False
-
-
-def remove_monitoring_stack_from_ocs(ocs_storage_classes):
-    monitoring_obj = ocp.OCP(
-        namespace=constants.MONITORING_NAMESPACE, kind='ConfigMap',
-    )
-    param_cmd = '[{"op": "replace", "path": "/data/config.yaml", "value": ""}]'
-    monitoring_obj.patch(
-        resource_name='cluster-monitoring-config',
-        params=param_cmd,
-        format_type='json'
-    )
